Complete word-count exercise
We make use of a hash to keep track of words we've seen and their counts. When we find one we haven't seen before, we add it with a count of one. In order to find them, we use `.gsub` with the regexp /[^a-z0-9']/ to replace any non alphanumeric or apostrophes with a space, then `.split` on spaces.